refactor: add EscapeHexCellWriteHandler for XLSX escape processing
- Moved EscapeHexCellWriteHandler to handler.impl package
- Registered EscapeHexCellWriteHandler in DefaultWriteHandlerLoader for XLSX format
- Updated XlsxEscapeUtils documentation to link to EscapeHexCellWriteHandler
- Improved handling of XLSX character escapes in cell writing process
